Move jwt verificaiton to http4s-util
