+++ app-backend/backsplash-server/src/main/resources/application.conf
@@ -31,7 +31,7 @@ cache {
   authorizationCacheEnable = ${?BACKSPLASH_AUTHORIZATION_CACHE_ENABLE}
 
   authenticationCacheEnable = false
-  authenticationCacheEnable = ${?BACKSPLASH_AUTHENTICATION_CACHE_ENABLE}
+  authenticationCacheEnable = ${?RF_AUTHENTICATION_CACHE_ENABLE}
 
   histogramCacheEnable = false
   histogramCacheEnable = ${?BACKSPLASH_HISTOGRAM_CACHE_ENABLE}
