+++ app-backend/http4s-util/src/main/resources/reference.conf
@@ -1,0 +1,4 @@
+cache {
+  authenticationCacheEnable = false
+  authenticationCacheEnable = ${?RF_AUTHENTICATION_CACHE_ENABLE}
+}
