+++ app-backend/http4s-util/src/main/scala/com/rasterfoundry/http4s/Config.scala
@@ -1,0 +1,13 @@
+package com.rasterfoundry.http4s
+
+import com.typesafe.config.ConfigFactory
+
+object Config {
+  private val config = ConfigFactory.load()
+
+  object cache {
+    private val cacheConfig = config.getConfig("cache")
+    val authenticationCacheEnable =
+      cacheConfig.getBoolean("authenticationCacheEnable")
+  }
+}
